Optimizied the tag profile page
